--- src/redux/admin/actions.js
@@ -1,6 +1,0 @@
-const actions = {
-  UPDATE_PROFILE: 'admin/UPDATE_PROFILE',
-  CHANGE_PASSWORD: 'admin/CHANGE_PASSWORD',
-}
-
-export default actions
