+++ src/redux/cart/actions.js
@@ -1,0 +1,3 @@
+const actions = {}
+
+export default actions
