+++ src/redux/cart/reducers.js
@@ -1,0 +1,12 @@
+import actions from './actions'
+
+const initialState = {}
+
+export default function cartReducer(state = initialState, action) {
+  switch (action.type) {
+    case actions.SET_STATE:
+      return { ...state, ...action.payload }
+    default:
+      return state
+  }
+}
