+++ src/redux/cart/sagas.js
@@ -1,0 +1,8 @@
+// import { all, takeEvery, put, putResolve, call } from 'redux-saga/effects'
+// import { notification } from 'antd'
+// import actions from './actions'
+
+export default function* rootSaga() {
+  // yield all([
+  // ])
+}
