--- src/redux/mentorships/mentorshipListing/actions.js
@@ -1,7 +1,0 @@
-const actions = {
-  CREATE_LISTING: 'mentorship/CREATE_LISTING',
-  UPDATE_LISTING: 'mentorship/UPDATE_LISTING',
-  DELETE_LISTING: 'mentorship/DELETE_LISTING',
-}
-
-export default actions
