+++ src/redux/reducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 import { connectRouter } from 'connected-react-router'
-import admin from './admin/reducers'
+import cart from './cart/reducers'
 import user from './user/reducers'
 import menu from './menu/reducers'
 import settings from './settings/reducers'
@@ -9,7 +9,7 @@ import categories from './categories/reducers'
 export default history =>
   combineReducers({
     router: connectRouter(history),
-    admin,
+    cart,
     user,
     menu,
     settings,
